Create collision shape builder
