--- Scenes/Test/PlanetTest.cs.uid
@@ -1,1 +1,0 @@
-uid://d4eqa33f81yrj
