--- Scenes/Test/PlanetTest.tscn
@@ -1,47 +1,0 @@
-[gd_scene load_steps=5 format=3 uid="uid://cfv1817f34ekw"]
-
-[ext_resource type="Script" uid="uid://x1l3aeqn7ho7" path="res://Entity/Planet/Planet.cs" id="1_2k4m3"]
-[ext_resource type="Script" uid="uid://d4eqa33f81yrj" path="res://Scenes/Test/PlanetTest.cs" id="1_awbyp"]
-
-[sub_resource type="SphereShape3D" id="SphereShape3D_fghij"]
-radius = 20.0
-
-[node name="PlanetTest" type="Node3D"]
-script = ExtResource("1_awbyp")
-planetNodePath = NodePath("Planet")
-
-[node name="DirectionalLight3D" type="DirectionalLight3D" parent="."]
-transform = Transform3D(1, 0, 0, 0, 0.866025, 0.5, 0, -0.5, 0.866025, 0, 10, 0)
-light_energy = 2.0
-shadow_enabled = true
-
-[node name="Camera3D" type="Camera3D" parent="."]
-transform = Transform3D(1, 0, 0, 0, 0.866025, 0.5, 0, -0.5, 0.866025, 0, 0, 10)
-current = true
-
-[node name="Planet" type="RigidBody3D" parent="." node_paths=PackedStringArray("_collisionShape")]
-script = ExtResource("1_2k4m3")
-SurfaceGravity = 9.8
-InitialMass = 1000.0
-GravitationalConstant = 0.1
-Radius = 20.0
-AutoCalculateOrbitalVelocity = false
-_collisionShape = NodePath("CollisionShape3D")
-Resolution = 256
-NoiseScale = 1.2
-NoiseOctaves = 7
-NoisePersistence = 0.25
-NoiseLacunarity = 3.0
-HeightScale = 0.1
-Roughness = 0.8
-Metallic = 0.0
-HasAtmosphere = true
-HasWater = true
-
-
-[node name="CollisionShape3D" type="CollisionShape3D" parent="Planet"]
-shape = SubResource("SphereShape3D_fghij")
-
-[node name="Timer" type="Timer" parent="Planet"]
-wait_time = 0.5
-autostart = true
